Change database schema for executor tags
Now the executor tags are stored as a JSON blob in the executors table and
the required tags for a flow are in the flow data blob in the execution_flows
table.

The checking of whether an executor can satisfy all of the required tags is
now done in Java code rather than in SQL, which means that logging can be
done to help diagnose executors that are not picking up executions.