Fix not saving group permissions when group name matches an existing user's name. (#2311)
User and group permissions are stored in the same database table with the 'name' column as the primary key. Groups are differentiated from users by having the 'isGroup' column set to true. Because the 'isGroup' column is not a primary key, attempts to create permissions for a group with the same name as an existing user end up updating the user of that name. To fix this issue is necessary to make 'isGroup' a primary key.