Don't delete (from HDFS) project files of old versions if shared with newer versions. (#2346)
The name of project files (zips) in HDFS is the hash of the file, this is good to avoid duplication of same files because if a user uploads a new project version that's the same as one uploaded previously it will share the project files with the latter. So, when deleting the project files of older versions, it's necessary to ensure the deleted files are not used by versions that are to be kept.