Don't allow pausing an execution after a KILL action has been committed.
The KILL action can take a while to be processed. During that waiting time the user was able to click the PAUSE button. If they did the execution would be marked as PAUSED even though all the running jobs would be killed anyway. In addition to persisting an incorrect flow status this causes that long running flows are not killed automatically and that executor server processes waiting for all executions to finish before stopping will be stuck.

In addition to the backend changes the PAUSE button will be disabled in the UI after the KILL button is clicked.