Fetch flow execution status from database column only. (#2474)
Currently flow execution status is stored in 2 places: a column in the DB and in the execution's JSON object. Because flow execution status is not read from the same source all the times sometimes we see inconsistency in the UI where different pages show different flow execution status.

This PR addresses that inconsistency by always reading status from the dedicated status DB column.