Rollback timezone feature in schedules. (#2519)
This feature was added in PR #2407 followed up with #2438 and #2507 which are being also rolled back here.
We're in the need to rollback this feature for now because it's causing too much confusion among users. The confusion is mainly due to:
-currently in Azkaban most of the times shown in the UI do not include the timezone.
-preview of 10 next runs in the schedule page is calculated based on the specified schedule timezone but after the schedule is created "Next Execution time" is shown in the server's timezone which doesn't necessarily match the preview.
Azkaban users do access and use the application from different timezones but some analysis is needed to come up with a solution that effectively helps and improves experience for those users.