Add thread names for multi-threaded pools
This makes it easier to identify the relevant pools during debugging.
For example - when looking at stack dumps and logs.
Notes:
(1) Explicitly avoids trying to name any pools created by libraries/frameworks
   azkaban uses. Such as Jetty (as they usually have their own conventions for this)
(2) Attaches the prefix `azk-` to clearly indicate a thread was created
   within Azkaban code
(3) While this is fairly exhaustive, it's possible that a few instances of new
   threads being created may have been missed.
   They can be addressed with subsequent changes as and when we identify those.

For reference, used this regex to locate relevant code `new\S+(Thread|Executor)`