Use new instance of HDFS object to fetch NameNode token as URI is unique (#2539)
Use new instance of HDFS object to fetch NameNode token as URI is unique with updated Kerberos Principal.

With older logic if there exists a FileSystem cache, the cache would blow up very quickly.
Earlier, the UGI for namenode used to be just the proxy user name. If a job with same user requested FileSytem object, it would be fetched from the cache.
With update to Kerberos principal to /az__<exec_id>, the UGI became unique to an execution thus blowing the FileSystem cache up.
Morever, the entry is never fetched from FileSystem cache as it is unique to that particular execution rendering the cache entry useless.
This PR alleviates this by creating a new instance of FS and closing it in the same method.