Fix DB connection leak in getValidationStatuses (#2550)
getValidationStatuses() gets a DB connection directly from the connection pool
instead of using DatabaseOperator.query. If a connection is fetched directly then
the method is also responsible for managing the lifecycle of the DB connection.

This change adds lifecycle management logic for Connection, ResultSet and
PreparedStatement objects.

Added a TODO to move to DatabaseOperator.query() since it takes care of the
lifecycle management for all these objects. This fix is time sensitive and
hence change to use query() will be done in a follow-up PR.

Tested the fix and verified no DB connection leak is present in getValidationStatuses().

Co-authored-by: Gaurav Salia <gsalia@gsalia-mn2.linkedin.biz>