Fix handling of Api failures during executor healthcheck
We have seen instances (in logs) where the ajax `ping` api used
for checking reachability to an executor can throw an unchecked
exception (NPE).
In the past this exception would have stopped the healthcheck
altogether but that has been fixed since 575c47.
This change makes the exception handling more granular and
makes healthchecking a lot more robust in case of unchecked
excecptions.