Remove direct instantiation and cast to Hdfs DistributedFileSystem
Current implememtation of HdfsStorage does a direct cast to
DistributedFileSystem(DFS), this was required because DFS provides a
public implementattion of rename method with the 'overwrite' option.
This method is protected in FileSystem.

For fixing this we replace the use of FileSystem with FileContext which
is the preferred way of using Hdfs and also deprecates many of the
methods in FileSystem.  However, it does not provide the
`copyFromLocalFile` method that FileSystem does and that functionality
is replaced with simple impelementation that directly uses Hdfs based
streams (quite similar to what what copyFromLocalFile does internally)

Also upgrades hadoop client version to 2.7.4

RB=2151659
BUG=LIHADOOP-52162,LIHADOOP-53795
G=azkabandev-reviewers
R=anath1,jakhani,djaiswal,soudas,vjalapar,ypadron,sshardoo,sarumuga,gsalia,adsharma,abtiwari,angoel,ygerchik,apruthi,prkhande,areznik,anismalh,chliang
A=djaiswal,anismalh,chliang