Support sla settings in execute & schedule API actions (#2523)
This implements "#2509 Support SLA options in execute API."

As a bonus SLA options are now supported also in the schedule call because it's using the same HttpRequestUtils. This is nice because if using the API directly to add schedules (like many users do), the API now allows ensuring that a scheduled execution doesn't start running before SLA has been set for it.

The separate set sla API for schedules remains as well.