Pass cluster object around instead of cluster id during job parameter construction in JobTypeManager.
The current two-step process was introduced to avoid passing Cluster objects around deep in method call chain.
This however, has made it cubersome to for ClusterRouter implementations to keep a consistent response to the
the same job, as cluster properties may change between the two steps for the same job. Specifically in Robin-
based Router, we'd like to allow Robin to override certain cluster properties and expose the merged view to
jobs. By merging the two steps together, internal bookkepping with Robin-based router can be avoided, simplifying
its implementation.