Fixed the thread-pool usage when downloading dependencies
Problem: when we download the dependencies of a project in thin-zip flow, we create a new thread-pool for every request. This creates a huge number of threads in the JVM which are Idle - observed as many as 1500+ threads in the ES process - all were from the same thread-pool dependency-transfer.
Fix: created a single executor process and reused across the calls.
Notes: In the previous version, if a single download fails we shutdown the ExecutorService to stop the currently running tasks. I have tried to keep the same behaviour by cancelling the future. However, cancellation will happen only if the task hasn't started yet. Added a config to adjust the thread-pool size if required. Refactored the unit tests.