[UI] Improve readability of long job property values
Job property values are presented in the UI as a single line of characters. This makes it difficult to read values such as: classpaths separated by comma or colon, list of options that are passed to a command/process.
To improve readability in such cases values following below patterns will be formatted in the UI by adding a new line where appropriate.
--param1 value1 --param2 --param3 value3
-param1 value -param2 -param3
something,something,something
something:something:something