Remove null from startup-dependencies.json while writing and reading … (#2588)
* Remove null from startup-dependencies.json while writing and reading the file.

A couple of users complained about not being able to execute flows. Debugging the
issue revealed that thinZip was enabled for this project and startup dependencies
that are saved in the DB contains a 'null'.

When 'null' is encountered in the dependencies json stored in the DB, an error occurs
while created Dependency object from the json since 'null' is not expected.

This change adds filters and checks to remove 'null' while writing and reading startup
dependencies. Also added unit test to ensure the code checks for null while writing
startup-dependencies.json to disk and while reading the dependencies json object
from the DB.

* Address review comments from PR to remove null in startup-dependencies.json

Co-authored-by: Gaurav Salia <gsalia@gsalia-mn2.linkedin.biz>

RB=2264841
BUG=LIHADOOP-55085
G=azkabandev-reviewers
R=areznik,prkhande,adsharma,soudas,abtiwari,sshardoo,ygerchik,sarumuga,ypadron,anath1,angoel,djaiswal,anismalh,gsalia,apruthi,jakhani
A=djaiswal