[UI] Improve readability of long job property values (#2580)
Job property values are presented in the UI as a single line of characters. This makes it difficult to read values such as: classpaths separated by comma or colon, list of options that are passed to a command/process.
To improve readability in such cases values following below patterns will be formatted in the UI by adding a new line where appropriate.
--param1 value1 --param2 --param3 value3
-param1 value -param2 -param3
something,something,something
something:something:something

RB=2284128
BUG=LIHADOOP-49659
G=azkabandev-reviewers
R=abtiwari,adsharma,ypadron,anath1,jakhani,angoel,areznik,soudas,sshardoo,anismalh,ygerchik,djaiswal,gsalia,apruthi,prkhande,sarumuga
A=djaiswal