Fix flaky queue duration condition in JobRunnerTest (#2604)
At least locally this test was often executed so fast that time in queue was set to 0. That must mean that System.currentTimeMillis() had not advanced between the times when job was queued and executed. Add 1ms sleep to ensure that queue time is always at least 1.

RB=2296887
G=azkabandev-reviewers
R=sarumuga,jakhani,anath1,anismalh,ygerchik,abtiwari,angoel,ypadron,areznik,gsalia,soudas,adsharma,sshardoo,apruthi,prkhande,djaiswal
A=djaiswal