Add TLS support to ExecutorApiGateway (#2679)
Modifies RestfulApiClient and ExecutorApiClient to use TLS for
connecting to https based servers.
- Introduces a protected method createHttpClient()
- The method is overridden in ExecutorApiClient to create a tls client
- Adds multiple sanity tests for the ExecutorApiClient
- Adds self-signed certificates, keys in a keystore and trustore for
  tests

Note that tls is not enabled on the jetty ExecutorServer as part of this
change. That support will be considered in future.  For now the
assumption is that an intervening proxy will provide the https support,
which will be the case with containerized deployments.

(cherry-picked from commit 7a18f7bdffffc2cfd53592fb8c3c95533986b8db)

PR Link: https://github.com/azkaban/azkaban/pull/2679

RB=2410096
G=azkabandev-reviewers
R=areznik,djaiswal,gsalia,angoel,anismalh,jakhani,ygerchik,ypadron,abtiwari,adsharma,sarumuga,anath1,apruthi,sshardoo,prkhande
A=sarumuga