Prevent kill action from leaving executions stuck in KILLING status. (#2691)
When a user requests to kill an execution a check that it's still running is made several times however the execution may finish before the kill action is processed after those checks causing it to stay in KILLING status forever. To prevent this add one last check in the same critical section where the kill is done.

RB=2423501
BUG=LIHADOOP-56882
G=azkabandev-reviewers
R=abtiwari,anath1,ypadron,ygerchik,gsalia,djaiswal,soudas,angoel,jakhani,prkhande,anismalh,sshardoo,areznik,apruthi,sarumuga,adsharma
A=gsalia