Don't expect Hadoop to always have the HDFS scheme configured. (#2803)
Previously, Azkaban always expected the HDFS scheme to be available in a Hadoop installation, so it would choose a default HDFS configuration when it didn't find the corresponding (HDFS specific) properties in the Hadoop configuration files. This is OK as long as a HDFS file system is available (possibly alongside others), but causes problems when it's not.

To fix this I'm removing the default Hadoop configurations provided by Azkaban, since, in practice, HDFS is usually and should be configured in Hadoop's config files. Although this is not a backward compatible change and issues may arise for those with Hadoop configurations not defining a file system properly this should be very rare.

Also, this change includes refactoring of the HadoopModule class (use of constants in the named bindings
and removal of Props object from constructor method) but no changes in functionality are expected from that.