Add listener for pod monitoring and deletion
This will ensure that flow-pods and corresponding clusterIp-services in
the 'completed' state will be deleted after flow execution finishes.
A subsequent commit will add this automated listener based deletion for
pods in 'errored' states as well.

Also, includes guice bindings for pod event listeners and a minor change
in ContainerizedDispatchManager for startup and shutdown of watch.