Allow longer project events to be recorded. (#2841)
User actions on projects are recorded in the DB and available for inspection.
Examples of those actions are uploading a new zip file, granting permissions to users and adding/modifying/deleting job properties in a flow through the UI.
When users update multiple job properties at a time or the value of one of them is very long it can happen (we often see it in the web server logs) that the event fails to be saved in the DB because the log message is too long.
To account for those scenarios the size of the MySQL column storing the log message is increased.

RB=2570239
G=azkabandev-reviewers
R=adsharma,angoel,sarumuga,apruthi,ygerchik,anismalh,prkhande,sshardoo,ypadron,anath1,areznik,gsalia,jakhani,abtiwari,djaiswal
A=jakhani