Add flow finalization and container cleaup invalid transtions
We observed cases where invalid `AzPodStatus` transitions would
result in corresponding Flows getting stuck in a non-final state
like Running.
This change expicitly adds finalization of flows for such cases
as well as deletion of kubernetes resources for such flows.
As an additional change, we now log the entire watch event for
cases where a flow had to be finalized, this will help determine
the root-cause and allow further analysis of failure scenarios.