The Image Management APIs allow multiple image versions of same type to remain in ACTIVE state. (#2952)
This leads to unexpected behavior of getting incorrect version among the ACTIVE ones for executing flows.
Introducing new image version state STABLE.
If an image is set as ACTIVE using the ACTIVATE API, all the other image versions which are currently
set to ACTIVE will be set to STABLE state.

RB=2858661
G=azkabandev-reviewers
R=sarumuga,adsharma,areznik,abtiwari,angoel,ypadron,ygerchik,sshardoo,apruthi,djaiswal,prkhande,jakhani,anismalh,anath1,gsalia
A=djaiswal