Make metrics reporting more open source friendly. (#2983)
Azkaban uses the Dropwizard Metrics library to collect metrics to gauge how the app is doing. To report those metrics for proper storage and plotting for example, it allows custom reporters to be hooked in.
Before, to connect a custom reporter the fully qualified name of its class had to be provided in a config property and the class should have a constructor with the following signature: (String reporterName, MetricRegistry registry, String metricsServerURL, Boolean someFlag);
This approach leaks LinkedIn specific settings for reporting and forces users to implement a constructor with parameters that might not make sense to them. In addition, it's not extensible as to accommodate new configs, in the past the constructor signature has been modified making it a backward incompatible change.

To make the configuration of custom reporters more open source friendly (able to support reporter specific settings and can withstand changes) a constructor with the following signature can be used instead: (MetricRegistry registry, String reporterConfigsPath)
-A MetricRegistry should always be required as it contains all the app metrics.
-Optionally, the path to a file containing custom reporter specific settings can be specified. The file format and how to read it is up to the reporter implementation.