Add ability to execute the Polling cleanup logic in Containerized dispatch model. (#3047)
The transition to CONTAINERIZED dispatch model from POLL model can be done gradually by splitting the workload between the 2. In the scenario where both dispatch models are in use both of the logics to cleanup stuck executions are needed.

This PR adds a workaround to run the ExecutorHealthChecker thread from POLL dispatch when Containerization is enabled. The ExecutorHealthChecker thread will start only if the properties azkaban.executor.healthcheck.interval.min and azkaban.executor.max.failurecount are configured in the azkaban.properties file.

A better solution would be to support multiple dispatch models in the web server instead of just one.