Fail azkaban flow when conditional variable cannot be resolved
Currently the conditional azkaban workflow will return null when
attempting to resolve a variable in the job properties. However,
this leads to strange exceptions such as ones seen here:

https://github.com/azkaban/azkaban/issues/1897

"javax.script.ScriptException: <eval>:1:1 Expected ; but found {"

This is not very helpful and indicitive of what the problem is.
This changeset fails hard if required variables cannot be found,
the job name cannot be found, or the properties don't exist.
It also provides helpful other information like the available properties
that have variables to choose from.