Sync Schedule with Trigger when web sever is up
Years ago, we separate schedule and trigger in favor to unify the use of trigger in the backend but aborted in the middle. It creates the sync issues between both. We have "TriggerBasedScheduleLoader" as the intermediate bridge to coordinate while trigger updates and ensure "schedule" updated to serve API request well and correctly.
This PR is to fix the missing sync on web server just restart and loading all the triggers into schedule format so schedule API can serve request immediately on restart, no need to wait for refresh to update them.