fix the bug that returned the outdated schedule metadata for getSchedule call (#3259)
For the legacy reason schedule and trigger are self-maintained and schedule itself maintains a cache to get trigger data more quickly especially when no update on trigger.
The better and cleaner fix is to remove the schedule layer but link API to trigger directly, in this way we don't have to maintain the status of schedule. However to quickly mitigate this issue, in this PR:

- fix the bug that store new schedule in cache;
- enable more debug level log to help increase debug efficiency.

RB=3879672
G=azkabandev-reviewers
R=gsalia,sshardoo,ypadron,abtiwari,anath1,apruthi,djaiswal,jakhani,adsharma,areznik,prkhande,sarumuga
A=rmnagara,amielke