fix flow trigger lost caused by failure project upload (#3285)
* fix flow trigger lost caused by failure project upload

Summary
Azkaban allows users to define flow triggers in HadoopDSL format, they're parsed and scheduled in the process of upload. The initial order is unscheduleTrigger -> handleUploadZip -> some postHandleWork such as rescheduleTrigger. However handleUploadZip step is a bit fragile and might fail due to intermittent issues from infra side, for example, artifatory down. It would lead to project upload in a broken stage where the previous triggers were removed by that broken upload.
This PR is to change the order of unscheduleTrigger, moving it after uploadZip.

Details
- adjust order of unschedule QuartzJob, so even if uploadFile fails, the previous trigger remains unimpacted.
- fix remove schedule logic by using getAllSchedule instead of getUpdatedSchedule.