Disable job property edit permission when project lock is set
Summary
This PR is the third step to address the security which restricted the permission to edit job properties when project lock is on. This limitation is aiming to protect production projects from any adhoc modifications unless the projects have to be gone through a review process and reuploaded by upload privileged user who intented to be set by a deployment automation user.
This restriction would also be put under a feature flag azkaban.disable.job.props.override.on.locked.
Additionally making a several improvements on overall UI settings.

Details
- Disable "Job Properties" Edit via UI and via API when project lock is set;
- Fix the lock status when adhocUpload restriction is not set, we should still properly change the project lock status;
- UI improvements:
  1. hide edit button for job properties when lock is on;
  2. show adhocUpload editable only when this feature flag is turned on; Normally we only show the status of project lock in UI;