Add support for schedule deletion notification
Add support to enable email notification for users when the schedule has been deleted. This is a default enabled feature for all flows, toggled by a feature flag. For email recipients, we first use SLA setting's email address, if the schedule does not have SLA set, we would look for success or failure email settings for executions as recipients.